+++ CHANGES.rst
@@ -6,11 +6,14 @@ Changes
 
 Improvements:
 
+- `spinta copy` for XSD supports globally defined attributes, referenced in other places.(`#605`_)
 - Updated `authlib` minimal version to 1.0.0 (`#675`_).
 - `private` and `public` keys now include `kid` field (`#675`_).
 
+  .. _#605: https://github.com/atviriduomenys/spinta/issues/605
   .. _#675: https://github.com/atviriduomenys/spinta/issues/675
 
+Other:
 - Removed dependency `mypy`
 
 0.2dev5 (2025-09-03)